refactor(security): rename package signed_token -> signedtoken
Drop underscore from package name to satisfy staticcheck ST1003 / Go
naming conventions.
